refactor: move Login and Signup components to pages directory and remove old files
