+++ src/pages/Signup.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Check, Eye, EyeOff, Lock, Mail, User } from 'lucide-react';
-import LoginButton from '../atoms/LoginButton';
-import EmailVerificationModal from '../modals/EmailVertificationModal';
+import LoginButton from '../components/atoms/LoginButton';
+import EmailVerificationModal from '../components/modals/EmailVertificationModal';
 import wallpaper from '../../assets/images/nubelson-fernandes-UcYBL5V0xWQ-unsplash.jpg';
 
 export default function Signup() {
